+++ src/emilybot/parser/__init__.py
@@ -1,4 +1,4 @@
-from emilybot.parser.command_parser import parse_command
+from emilybot.parser.command_parser import parse_message
 from emilybot.parser.types import JS, Command, ListChildren
 
-__all__ = ["parse_command", "Command", "JS", "ListChildren"]
+__all__ = ["parse_message", "Command", "JS", "ListChildren"]